Changed node and edge indices from ints to longs so that now indices have the same value as the SUID.

+++ model-api/src/main/java/org/cytoscape/model/CyEdge.java
@@ -89,7 +89,7 @@ public interface CyEdge extends CyIdentifiable {
 	 * necessarily an index into {@link CyNetwork#getEdgeList}.
 	 * @return An index for this edge within this network.
 	 */
-	int getIndex();
+	long getIndex();
 
 	/**
 	 * Returns the source node determining this edge. It will never be null.
